修复多线程安全问题、性能优化与健壮性提升，新增详细设计文档
Bug 修复：
- FaceRecognitionFilter: 添加 mutex 保护 frameCounter_/cachedIdentities_ 数据竞争
- FaceTracker: nextTrackId_ 改为 std::atomic<int> 避免数据竞争
- OpenCvDnnDetector: SSD 输出张量访问前添加列数边界检查
- AttendanceObserver: 修复 localtime() 空指针风险，添加文件打开失败日志

性能优化：
- PreprocessFilter: cv::Mat 赋值改用 std::move 避免不必要拷贝
- FaceDatabase: 构造时缓存识别阈值，避免每次 match() 重复读取配置
- FaceDetectionFilter: 边界框缩放使用 float 运算避免整数精度损失

健壮性提升：
- CudaResourceManager: 检查 cudaGetDeviceProperties() 返回值
- ConfigManager: 捕获具体 JSON 解析错误并记录详细信息
- PipeProcessor: 消除 write() 未使用返回值警告
- server.py: 将阻塞 select 调用移入 run_in_executor 避免阻塞事件循环
- webrtc-client.js: 添加 Image onerror 处理释放 Object URL 防止内存泄漏
- ui.js: 修复 FPS 计数器重复计数问题

文档：
- 新增 docs/design/software_architecture.md（软件体系结构分析，10 章）
- 新增 docs/design/design_patterns.md（设计模式分析，10 章）
- 更新 README.md 文档引用和项目结构

+++ src/pipeline/filters/FaceDetectionFilter.cpp
@@ -53,10 +53,10 @@ bool FaceDetectionFilter::process(VideoFrame& frame) {
     if (scale < 1.0f) {
         float invScale = 1.0f / scale;
         for (auto& face : frame.detectedFaces) {
-            face.bbox.x = static_cast<int>(face.bbox.x * invScale);
-            face.bbox.y = static_cast<int>(face.bbox.y * invScale);
-            face.bbox.width = static_cast<int>(face.bbox.width * invScale);
-            face.bbox.height = static_cast<int>(face.bbox.height * invScale);
+            face.bbox.x = static_cast<int>(static_cast<float>(face.bbox.x) * invScale);
+            face.bbox.y = static_cast<int>(static_cast<float>(face.bbox.y) * invScale);
+            face.bbox.width = static_cast<int>(static_cast<float>(face.bbox.width) * invScale);
+            face.bbox.height = static_cast<int>(static_cast<float>(face.bbox.height) * invScale);
         }
     }
